Re-allocated python modules
