+++ python/sparknlp/__init__.py
@@ -7,9 +7,6 @@ sys.modules['com.johnsnowlabs.nlp.annotators.ner'] = annotator
 sys.modules['com.johnsnowlabs.nlp.annotators.ner.regex'] = annotator
 sys.modules['com.johnsnowlabs.nlp.annotators.ner.crf'] = annotator
 sys.modules['com.johnsnowlabs.nlp.annotators.ner.dl'] = annotator
-sys.modules['com.johnsnowlabs.nlp.annotators.assertion'] = annotator
-sys.modules['com.johnsnowlabs.nlp.annotators.assertion.logreg'] = annotator
-sys.modules['com.johnsnowlabs.nlp.annotators.assertion.dl'] = annotator
 sys.modules['com.johnsnowlabs.nlp.annotators.pos'] = annotator
 sys.modules['com.johnsnowlabs.nlp.annotators.pos.perceptron'] = annotator
 sys.modules['com.johnsnowlabs.nlp.annotators.sbd'] = annotator
