+++ python/sparknlp/common.py
@@ -1,19 +1,85 @@
-import sparknlp.internal as _internal
+from pyspark.ml.util import JavaMLWritable
+from pyspark.ml.wrapper import JavaModel, JavaEstimator
+from pyspark.ml.param.shared import Param, TypeConverters
+from sparknlp.util import AnnotatorJavaMLReadable
 from pyspark.ml.param import Params
+from pyspark import keyword_only
+import sparknlp.internal as _internal
 import re
 
 
-def RegexRule(rule, identifier):
-    return _internal._RegexRule(rule, identifier).apply()
+class AnnotatorProperties(Params):
+
+    inputCols = Param(Params._dummy(),
+                      "inputCols",
+                      "previous annotations columns, if renamed",
+                      typeConverter=TypeConverters.toListString)
+    outputCol = Param(Params._dummy(),
+                      "outputCol",
+                      "output annotation column. can be left default.",
+                      typeConverter=TypeConverters.toString)
+
+    def setInputCols(self, value):
+        return self._set(inputCols=value)
+
+    def setOutputCol(self, value):
+        return self._set(outputCol=value)
 
 
-class ReadAs(object):
-    LINE_BY_LINE = "LINE_BY_LINE"
-    SPARK_DATASET = "SPARK_DATASET"
+class ApproachWithEmbeddings(Params):
+    embeddingsDim = Param(Params._dummy(),
+                          "embeddingsDim",
+                          "Number of embedding dimensions",
+                          typeConverter=TypeConverters.toInt)
 
+    caseSensitiveEmbeddings = Param(Params._dummy(),
+                                    "caseSensitiveEmbeddings",
+                                    "whether to ignore case in tokens for embeddings matching",
+                                    typeConverter=TypeConverters.toBoolean)
 
-def ExternalResource(path, read_as=ReadAs.LINE_BY_LINE, options={}):
-    return _internal._ExternalResource(path, read_as, options).apply()
+    includeEmbeddings = Param(Params._dummy(),
+                              "includeEmbeddings",
+                              "whether to include embeddings when saving annotator",
+                              typeConverter=TypeConverters.toBoolean)
+
+    embeddingsRef = Param(Params._dummy(),
+                          "embeddingsRef",
+                          "if sourceEmbeddingsPath was provided, name them with this ref. Otherwise, use embeddings by this ref",
+                          typeConverter=TypeConverters.toString)
+
+    sourceEmbeddingsPath = Param(Params._dummy(),
+                                 "sourceEmbeddingsPath",
+                                 "Word embeddings file",
+                                 typeConverter=TypeConverters.toString)
+    embeddingsFormat = Param(Params._dummy(),
+                             "embeddingsFormat",
+                             "Word vectors file format",
+                             typeConverter=TypeConverters.toInt)
+
+    def setEmbeddingsSource(self, path, nDims, format):
+        self._set(sourceEmbeddingsPath=path)
+        self._set(embeddingsFormat=format)
+        return self._set(embeddingsDim=nDims)
+
+    def setEmbeddingsDim(self, value):
+        return self._set(embeddingsDim=value)
+
+    def setCaseSensitiveEmbeddings(self, value):
+        return self._set(caseSensitiveEmbeddings=value)
+
+    def setIncludeEmbeddings(self, value):
+        return self._set(includeEmbeddings=value)
+
+    def setEmbeddingsRef(self, value):
+        return self._set(embeddingsRef=value)
+
+    def __init__(self):
+        super(ApproachWithEmbeddings, self).__init__()
+        self._setDefault(
+            caseSensitiveEmbeddings=False,
+            includeEmbeddings=True,
+            embeddingsRef=self.uid
+        )
 
 
 # Helper class used to generate the getters for all params
@@ -49,3 +115,74 @@ class ParamsGettersSetters(Params):
             self.set(self.getParam(paramName), v)
             return self
         return r
+
+
+class AnnotatorModel(JavaModel, AnnotatorJavaMLReadable, JavaMLWritable, AnnotatorProperties, ParamsGettersSetters):
+
+    column_type = "array<struct<annotatorType:string,begin:int,end:int,metadata:map<string,string>>>"
+
+    @keyword_only
+    def setParams(self):
+        kwargs = self._input_kwargs
+        return self._set(**kwargs)
+
+    @keyword_only
+    def __init__(self, classname, java_model=None):
+        super(AnnotatorModel, self).__init__(java_model=java_model)
+        if classname and not java_model:
+            self.__class__._java_class_name = classname
+            self._java_obj = self._new_java_obj(classname, self.uid)
+        if java_model is not None:
+            self._transfer_params_from_java()
+
+
+class ModelWithEmbeddings(AnnotatorModel):
+
+    embeddingsDim = Param(Params._dummy(),
+                          "embeddingsDim",
+                          "Number of embedding dimensions",
+                          typeConverter=TypeConverters.toInt)
+
+    caseSensitiveEmbeddings = Param(Params._dummy(),
+                                    "caseSensitiveEmbeddings",
+                                    "whether to ignore case in tokens for embeddings matching",
+                                    typeConverter=TypeConverters.toBoolean)
+
+    includeEmbeddings = Param(Params._dummy(),
+                              "includeEmbeddings",
+                              "whether to include embeddings when saving annotator",
+                              typeConverter=TypeConverters.toBoolean)
+
+    embeddingsRef = Param(Params._dummy(),
+                          "embeddingsRef",
+                          "if sourceEmbeddingsPath was provided, name them with this ref. Otherwise, use embeddings by this ref",
+                          typeConverter=TypeConverters.toString)
+
+    @keyword_only
+    def __init__(self, classname, java_model=None):
+        super(ModelWithEmbeddings, self).__init__(classname=classname, java_model=java_model)
+
+    def getClusterEmbeddings(self):
+        return self._java_obj.getClusterEmbeddings()
+
+
+class AnnotatorApproach(JavaEstimator, JavaMLWritable, AnnotatorJavaMLReadable, AnnotatorProperties,
+                        ParamsGettersSetters):
+    @keyword_only
+    def __init__(self, classname):
+        ParamsGettersSetters.__init__(self)
+        self.__class__._java_class_name = classname
+        self._java_obj = self._new_java_obj(classname, self.uid)
+
+
+def RegexRule(rule, identifier):
+    return _internal._RegexRule(rule, identifier).apply()
+
+
+class ReadAs(object):
+    LINE_BY_LINE = "LINE_BY_LINE"
+    SPARK_DATASET = "SPARK_DATASET"
+
+
+def ExternalResource(path, read_as=ReadAs.LINE_BY_LINE, options={}):
+    return _internal._ExternalResource(path, read_as, options).apply()
